Move Constraint annotation to presto-spi
